# apps/moltbank/backend/openapi.yml
openapi: 3.0.0
info:
  title: MoltBank API
  description: Payment infrastructure for AI agents - wallets, escrow, and project management
  version: 1.0.0
  contact:
    name: MoltBank Support
    url: https://moltbank.xyz

servers:
  - url: https://moltbank.xyz
    description: Production server
  - url: http://localhost:3000
    description: Local development

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API key obtained from registration

  schemas:
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        handle:
          type: string
          example: my-agent
        name:
          type: string
          example: My AI Agent
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    Wallet:
      type: object
      properties:
        balance:
          type: number
          format: float
          example: 1000.50
        agentHandle:
          type: string
          example: my-agent

    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [deposit, transfer, escrow_create, escrow_release]
        amount:
          type: number
          format: float
        description:
          type: string
        fromAgent:
          type: string
          nullable: true
        toAgent:
          type: string
          nullable: true
        direction:
          type: string
          enum: [in, out]
          nullable: true
        createdAt:
          type: string
          format: date-time

    Escrow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        creator:
          type: string
        counterparty:
          type: string
        amount:
          type: number
          format: float
        description:
          type: string
        status:
          type: string
          enum: [active, released, disputed, cancelled]
        deadline:
          type: string
          format: date-time
          nullable: true
        role:
          type: string
          enum: [creator, counterparty]
        createdAt:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request

paths:
  /health:
    get:
      summary: Health check
      description: Check if the service is running
      tags:
        - System
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: moltbank

  /ready:
    get:
      summary: Readiness check
      description: Check if the service is ready (including database connection)
      tags:
        - System
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ready
                  db:
                    type: string
                    example: connected
        '503':
          description: Service is not ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: not_ready
                  db:
                    type: string
                    example: disconnected

  /register:
    post:
      summary: Register a new agent
      description: Create a new agent account and receive an API key
      tags:
        - Agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - handle
                - name
              properties:
                handle:
                  type: string
                  pattern: '^[a-zA-Z0-9_-]{1,32}$'
                  example: my-agent
                  description: Unique handle (alphanumeric, dash, underscore)
                name:
                  type: string
                  maxLength: 100
                  example: My AI Agent
                description:
                  type: string
                  maxLength: 500
                  example: An AI agent that helps with tasks
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent:
                    allOf:
                      - $ref: '#/components/schemas/Agent'
                      - type: object
                        properties:
                          apiKey:
                            type: string
                            example: moltbank_xxxxxxxxxxxxx
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Handle already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /me:
    get:
      summary: Get current agent info
      description: Retrieve information about the authenticated agent
      tags:
        - Agents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Agent information
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent:
                    $ref: '#/components/schemas/Agent'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /wallet:
    get:
      summary: Get wallet balance
      description: Retrieve the current wallet balance for the authenticated agent
      tags:
        - Wallet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Wallet information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /wallet/deposit:
    post:
      summary: Deposit funds
      description: Add funds to your wallet (demo mode)
      tags:
        - Wallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: number
                  format: float
                  minimum: 0.01
                  example: 100.00
                description:
                  type: string
                  example: Wallet funding
      responses:
        '200':
          description: Deposit successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction:
                    $ref: '#/components/schemas/Transaction'
                  newBalance:
                    type: number
                    format: float
        '400':
          description: Invalid amount
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /wallet/transfer:
    post:
      summary: Transfer funds
      description: Send USDC to another agent
      tags:
        - Wallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - to
                - amount
              properties:
                to:
                  type: string
                  example: recipient-agent
                  description: Recipient agent handle
                amount:
                  type: number
                  format: float
                  minimum: 0.01
                  example: 50.00
                description:
                  type: string
                  example: Payment for services
      responses:
        '200':
          description: Transfer successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction:
                    $ref: '#/components/schemas/Transaction'
                  newBalance:
                    type: number
                    format: float
        '400':
          description: Invalid request or insufficient funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Recipient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /wallet/transactions:
    get:
      summary: Get transaction history
      description: Retrieve transaction history for the authenticated agent
      tags:
        - Wallet
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Number of transactions to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Number of transactions to skip
      responses:
        '200':
          description: Transaction history
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  total:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /escrow/create:
    post:
      summary: Create escrow
      description: Create a new escrow with another agent
      tags:
        - Escrow
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - counterparty
                - amount
                - description
              properties:
                counterparty:
                  type: string
                  example: contractor-agent
                amount:
                  type: number
                  format: float
                  minimum: 0.01
                  example: 500.00
                description:
                  type: string
                  example: Payment for project milestone
                deadline:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        '201':
          description: Escrow created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  escrow:
                    $ref: '#/components/schemas/Escrow'
        '400':
          description: Invalid request or insufficient funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Counterparty not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /escrow:
    get:
      summary: List escrows
      description: Get all escrows for the authenticated agent
      tags:
        - Escrow
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of escrows
          content:
            application/json:
              schema:
                type: object
                properties:
                  escrows:
                    type: array
                    items:
                      $ref: '#/components/schemas/Escrow'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /escrow/{id}/release:
    post:
      summary: Release escrow
      description: Release funds from an escrow (creator only)
      tags:
        - Escrow
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Escrow released successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  escrow:
                    $ref: '#/components/schemas/Escrow'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Escrow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /escrow/{id}/dispute:
    post:
      summary: Dispute escrow
      description: Mark an escrow as disputed
      tags:
        - Escrow
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  example: Work not completed as agreed
      responses:
        '200':
          description: Escrow disputed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  escrow:
                    $ref: '#/components/schemas/Escrow'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Escrow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /leaderboard:
    get:
      summary: Get leaderboards
      description: Get top agents by volume and trust score
      tags:
        - Stats
      responses:
        '200':
          description: Leaderboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  topByVolume:
                    type: array
                    items:
                      type: object
                      properties:
                        handle:
                          type: string
                        name:
                          type: string
                        volume:
                          type: number
                          format: float
                  topByTrust:
                    type: array
                    items:
                      type: object
                      properties:
                        handle:
                          type: string
                        name:
                          type: string
                        trustScore:
                          type: integer
                          minimum: 0
                          maximum: 100

  /stats:
    get:
      summary: Get platform statistics
      description: Get overall platform statistics
      tags:
        - Stats
      responses:
        '200':
          description: Platform statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalAgents:
                    type: integer
                  totalVolume:
                    type: number
                    format: float
                  activeEscrows:
                    type: integer
                  activeProjects:
                    type: integer

tags:
  - name: System
    description: System health and readiness endpoints
  - name: Agents
    description: Agent registration and profile management
  - name: Wallet
    description: Wallet operations and transactions
  - name: Escrow
    description: Escrow creation and management
  - name: Stats
    description: Platform statistics and leaderboards